Update rps.js
Go ahead and reprimand me if you don't appreciate my prying. I was looking for something else on Firefox add-ons and stumbled across this by mistake and thought "What the heck" and gave it a shot. I'll admit I suck at coding myself, but when it comes to debugging when I'm bored I find the small details ;). It's somewhat entertaining for a rainy day. Nice job though on getting most of the hard stuff done.